Added temporary alert on announcement page
added announcement data, reconfigured services directories

added a bit of styling, added created dates to announcements

added radio button to announcements, refactoring

responsive style added to sie menu bar

adjusted initial announcement data, added unit test

test refactor

refactoring announcement alert logic

Announcement refactoring WIP

added capability to change alert splash message

Added capability to add and delete announcements

refactor

+++ rest/src/main/java/com/constructionprojectmanager/rest/announcement/AnnouncementRepository.java
@@ -9,5 +9,5 @@ import javax.persistence.Table;
 @RepositoryRestResource
 @CrossOrigin(origins = "http://localhost:4200")
 @Table(name="announcement")
-public interface AnnouncementRepository  extends CrudRepository<Announcement, Long> {
+public interface AnnouncementRepository  extends CrudRepository<Announcement, Integer> {
 }
